pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cybercqbench"
version = "1.0.0"
description = "Research-grade evaluation platform for SOC/GRC LLM performance"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CyberCQBench Team"},
]
keywords = ["llm", "evaluation", "security", "soc", "grc", "cybersecurity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0", 
    "motor>=3.3.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2",
    "tiktoken>=0.5.2",
    "ulid-py>=1.1.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.3",
    "scipy>=1.11.4",
    "openai>=1.3.8",
    "anthropic>=0.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1", 
    "pytest-httpx>=0.28.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "mypy>=1.7.1",
]

[project.urls]
Homepage = "https://github.com/your-org/cybercqbench"
Documentation = "https://github.com/your-org/cybercqbench/docs"
Repository = "https://github.com/your-org/cybercqbench"
Issues = "https://github.com/your-org/cybercqbench/issues"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "B", "A", "COM", "DTZ", "EM", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "S101", "S106", "PLR0913", "PLR0912", "PLR0915", "TRY003"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "PLR2004"]
"scripts/*" = ["T201"]

[tool.black]
target-version = ['py311']
line-length = 100
skip-string-normalization = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "ulid.*"
ignore_missing_imports = true

[[tool.mypy.overrides]] 
module = "tiktoken.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
